Messing around with a small refactor of the tag parser
