+++ lib/bridgetown-inline-svg.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
 require "bridgetown"
+require "bridgetown-inline-svg/svg-tag"
 
 module BridgetownInlineSvg
-  autoload :SvgTag, "bridgetown-inline-svg/svg-tag"
 end
-
-Liquid::Template.register_tag("svg", BridgetownInlineSvg::SvgTag)
