+++ spec/fixtures/src/index.html
@@ -12,6 +12,7 @@ svgclass: hello
 <div id="height">
 {% svg images/square.svg  width=24 %}
 {% svg images/square.svg  width=24 height=48 %}
+{% svg images/square.svg width: 24 height: 48 %}
 {% svg images/square.svg  width=24 height="" %}
 </div>
 <div id="attributes">
